+++ src/nebulae.cpp
@@ -340,8 +340,10 @@ struct Nebulae : SanguineModule {
 				   between the input and output SRC.
 				*/
 				for (int block = 0; block < outputLength; ++block) {
-					input[block].l = clamp(convertedFrames[block].samples[currentChannel] * 32767.f, -32768.f, 32767.f);
-					input[block].r = clamp(convertedFrames[block].samples[currentChannel + 1] * 32767.f, -32768.f, 32767.f);
+					input[block].l = static_cast<short>(
+						clamp(convertedFrames[block].samples[currentChannel] * 32767.f, -32768.f, 32767.f));
+					input[block].r = static_cast<short>(
+						clamp(convertedFrames[block].samples[currentChannel + 1] * 32767.f, -32768.f, 32767.f));
 				}
 
 				cloudsParameters[channel] = cloudsProcessors[channel]->mutable_parameters();
@@ -424,8 +426,8 @@ struct Nebulae : SanguineModule {
 
 				// Convert output buffer.
 				for (size_t block = 0; block < clouds::kMaxBlockSize; ++block) {
-					renderedFrames[block].samples[currentChannel] = output[block].l / 32768.f;
-					renderedFrames[block].samples[currentChannel + 1] = output[block].r / 32768.f;
+					renderedFrames[block].samples[currentChannel] = static_cast<float>(output[block].l / 32768.f);
+					renderedFrames[block].samples[currentChannel + 1] = static_cast<float>(output[block].r / 32768.f);
 				}
 			}
